DAOS-12392 test: Allow test_core_files to pass in launch.py (#11231)
A change was made to fail launch.py if any test passed but created a core file.
This change also had the side effect of failing launch if test_core_files ran
correctly (i.e. created and processed a core file). This PR will allow launch.py
to PASS if test_core_files* correctly creates and processes a corefile.
