+++ src/tests/ftest/harness/core_files.yaml
@@ -1,0 +1,15 @@
+hosts:
+  test_servers: 2
+timeout: 120
+server_config:
+  engines_per_host: 1
+  engines:
+    0:
+      storage:
+        0:
+          class: ram
+          scm_mount: /mnt/daos
+          scm_size: 16
+pool:
+  size: 2G
+  control_method: dmg
